【优化】创建主表索引过程中，子表索引创建失败仍然继续创建其他子表索引。 删除主表索引时，如果某个子表存在该索引则认为删除成功，如果所有子表都不存在该索引时，则返回索引不存在
git-svn-id: http://192.168.20.11/sequoiadb/trunk@10337 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsShardSession.cpp
@@ -1982,12 +1982,19 @@ namespace engine
       iter = strSubCLList.begin();
       while( iter != strSubCLList.end() )
       {
-         rc = rtnCreateIndexCommand( iter->c_str(), boIndex, _pEDUCB,
+         INT32 rcTmp = SDB_OK;
+         rcTmp = rtnCreateIndexCommand( iter->c_str(), boIndex, _pEDUCB,
                                      _pDmsCB, _pDpsCB );
-         PD_CHECK( SDB_OK == rc || SDB_IXM_REDEF == rc, rc, error, PDERROR,
+         if ( SDB_OK != rcTmp && SDB_IXM_REDEF != rcTmp )
+         {
+            PD_LOG( PDERROR,
                    "create index for sub-collection(%s) failed(rc=%d)",
-                   iter->c_str(), rc );
-         rc = SDB_OK;
+                   iter->c_str(), rcTmp );
+            if ( SDB_OK == rc )
+            {
+               rc = rcTmp;
+            }
+         }
          ++iter;
       }
    done:
@@ -2009,6 +2016,7 @@ namespace engine
       set< std::string > strSubCLList;
       set< std::string >::iterator iter;
       BSONElement ele;
+      BOOLEAN isExist = FALSE;
       try
       {
          boMatcher = BSONObj( pQuery );
@@ -2031,13 +2039,28 @@ namespace engine
       iter = strSubCLList.begin();
       while( iter != strSubCLList.end() )
       {
-         rc = rtnDropIndexCommand( iter->c_str(), ele, _pEDUCB,
-                                   _pDmsCB, _pDpsCB );
-         PD_CHECK( SDB_OK == rc || SDB_IXM_NOTEXIST == rc, rc, error, PDERROR,
+         INT32 rcTmp = SDB_OK;
+         rcTmp = rtnDropIndexCommand( iter->c_str(), ele, _pEDUCB,
+                                    _pDmsCB, _pDpsCB );
+         if ( SDB_OK == rcTmp )
+         {
+            isExist = TRUE;
+         }
+         else
+         {
+            if ( SDB_OK == rc || SDB_IXM_NOTEXIST == rc )
+            {
+               rc = rcTmp;
+            }
+            PD_LOG( PDERROR,
                    "drop index for sub-collection(%s) failed(rc=%d)",
-                   iter->c_str(), rc );
-         rc = SDB_OK;
+                   iter->c_str(), rcTmp );
+         }
          ++iter;
+      }
+      if ( SDB_IXM_NOTEXIST == rc && isExist )
+      {
+         rc = SDB_OK;
       }
    done:
       return rc;
